Add visual graph editor.
This is just a first commit for @typemytype and @LettError to look at for functionality, design aesthetics, code style and terminology. The parsing and dumping will be moved to the appropriate files, the imports will move to the top of the file, and so on. I'm just keeping it in one blob for easier review for now.